requirements.txt
pyrogram==2.0.29
pytz==2023.3
python-dotenv==1.0.0
humanize==4.6.0
retrying==1.3.4
pyrate-limiter==3.3.2
rarfile==4.0
py7zr==0.21.2
pyzipper==0.6.1
tqdm==4.65.0
pyfiglet==0.8.post1
termcolor==2.3.0
chardet==5.1.0
slugify==5.0.2